refactor builder.go
